Moved application command managers, added declaration interfaces
