+++ src/main/kotlin/io/github/freya022/botcommands/api/commands/annotations/Command.kt
@@ -1,8 +1,9 @@
 package io.github.freya022.botcommands.api.commands.annotations
 
-import io.github.freya022.botcommands.api.commands.application.annotations.AppDeclaration
 import io.github.freya022.botcommands.api.commands.application.context.annotations.JDAMessageCommand
 import io.github.freya022.botcommands.api.commands.application.context.annotations.JDAUserCommand
+import io.github.freya022.botcommands.api.commands.application.declaration.GlobalApplicationCommandsDeclaration
+import io.github.freya022.botcommands.api.commands.application.declaration.GuildApplicationCommandsDeclaration
 import io.github.freya022.botcommands.api.commands.application.slash.annotations.JDASlashCommand
 import io.github.freya022.botcommands.api.commands.text.annotations.JDATextCommandVariation
 import io.github.freya022.botcommands.api.commands.text.annotations.TextDeclaration
@@ -17,7 +18,8 @@ import io.github.freya022.botcommands.api.core.service.annotations.BService
  *
  * @see BService @BService
  *
- * @see AppDeclaration @AppDeclaration
+ * @see GlobalApplicationCommandsDeclaration Declaring global application commands using the DSL
+ * @see GuildApplicationCommandsDeclaration Declaring guild application commands using the DSL
  * @see JDASlashCommand @JDASlashCommand
  * @see JDAMessageCommand @JDAMessageCommand
  * @see JDAUserCommand @JDAUserCommand
