+++ src/main/kotlin/io/github/freya022/botcommands/api/commands/application/context/annotations/JDAUserCommand.kt
@@ -1,13 +1,14 @@
 package io.github.freya022.botcommands.api.commands.application.context.annotations
 
 import io.github.freya022.botcommands.api.commands.annotations.*
-import io.github.freya022.botcommands.api.commands.application.AbstractApplicationCommandManager
 import io.github.freya022.botcommands.api.commands.application.ApplicationCommand
 import io.github.freya022.botcommands.api.commands.application.CommandScope
-import io.github.freya022.botcommands.api.commands.application.annotations.AppDeclaration
 import io.github.freya022.botcommands.api.commands.application.context.builder.UserCommandBuilder
 import io.github.freya022.botcommands.api.commands.application.context.user.GlobalUserEvent
 import io.github.freya022.botcommands.api.commands.application.context.user.GuildUserEvent
+import io.github.freya022.botcommands.api.commands.application.declaration.AbstractApplicationCommandManager
+import io.github.freya022.botcommands.api.commands.application.declaration.GlobalApplicationCommandsDeclaration
+import io.github.freya022.botcommands.api.commands.application.declaration.GuildApplicationCommandsDeclaration
 import io.github.freya022.botcommands.api.core.entities.InputUser
 import io.github.freya022.botcommands.api.localization.annotations.LocalizationBundle
 import io.github.freya022.botcommands.api.localization.context.AppLocalizationContext
@@ -45,7 +46,8 @@ import net.dv8tion.jda.api.interactions.commands.localization.LocalizationFuncti
  * @see RateLimit @RateLimit
  * @see Filter @Filter
  *
- * @see AppDeclaration Declaring application commands using the DSL
+ * @see GlobalApplicationCommandsDeclaration Declaring global application commands using the DSL
+ * @see GuildApplicationCommandsDeclaration Declaring guild application commands using the DSL
  * @see AbstractApplicationCommandManager.userCommand DSL equivalent
  */
 @Target(AnnotationTarget.FUNCTION)
