+++ src/main/kotlin/io/github/freya022/botcommands/api/commands/application/declaration/AbstractApplicationCommandManager.kt
@@ -1,6 +1,7 @@
-package io.github.freya022.botcommands.api.commands.application
+package io.github.freya022.botcommands.api.commands.application.declaration
 
 import io.github.freya022.botcommands.api.commands.annotations.Command
+import io.github.freya022.botcommands.api.commands.application.CommandScope
 import io.github.freya022.botcommands.api.commands.application.context.annotations.ContextOption
 import io.github.freya022.botcommands.api.commands.application.context.annotations.JDAMessageCommand
 import io.github.freya022.botcommands.api.commands.application.context.annotations.JDAUserCommand
