+++ src/main/kotlin/io/github/freya022/botcommands/api/commands/application/slash/annotations/JDASlashCommand.kt
@@ -1,10 +1,11 @@
 package io.github.freya022.botcommands.api.commands.application.slash.annotations
 
 import io.github.freya022.botcommands.api.commands.annotations.*
-import io.github.freya022.botcommands.api.commands.application.AbstractApplicationCommandManager
 import io.github.freya022.botcommands.api.commands.application.ApplicationCommand
 import io.github.freya022.botcommands.api.commands.application.CommandScope
-import io.github.freya022.botcommands.api.commands.application.annotations.AppDeclaration
+import io.github.freya022.botcommands.api.commands.application.declaration.AbstractApplicationCommandManager
+import io.github.freya022.botcommands.api.commands.application.declaration.GlobalApplicationCommandsDeclaration
+import io.github.freya022.botcommands.api.commands.application.declaration.GuildApplicationCommandsDeclaration
 import io.github.freya022.botcommands.api.commands.application.slash.GlobalSlashEvent
 import io.github.freya022.botcommands.api.commands.application.slash.GuildSlashEvent
 import io.github.freya022.botcommands.api.commands.application.slash.builder.SlashCommandBuilder
@@ -53,7 +54,8 @@ import net.dv8tion.jda.api.interactions.commands.localization.LocalizationFuncti
  * @see RateLimit @RateLimit
  * @see Filter @Filter
  *
- * @see AppDeclaration Declaring application commands using the DSL
+ * @see GlobalApplicationCommandsDeclaration Declaring global application commands using the DSL
+ * @see GuildApplicationCommandsDeclaration Declaring guild application commands using the DSL
  * @see AbstractApplicationCommandManager.slashCommand DSL equivalent
  */
 @Target(AnnotationTarget.FUNCTION)
